Extract libyear calculator
